make.conf: switch installed documentation to gzip
Since pigz runs faster than bzip2, and the installed documentation
is not that big (on the order of <500 MB compressed across all of
our packages), switch to gzip compression here.  This will increase
the installed footprint by <10 MB, but speed up (re)builds.

Ideally we'd use zstd here so it'd be smaller & faster, but man-db
does not understand zstd-compressed man-pages, so we have to stick
to the algorithms it supports.

BUG=b:187789829
TEST=CQ passes

Change-Id: I6b55fc35dd56cdfa3757a139101980bc569d6389
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3573755
Reviewed-by: Sloan Johnson <sloanjohnson@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>