dev-go/xxhash: don't build the xxhsum binary
We only need the module source packages as a dependency
during the Go modules conversion an the binary is not
called at all, it was just compiled for completeness.

This binary however conflicts with a C implementation
of xxhash [1] which also need to be added, so we drop
it to conserve some disk space.

If necessary, it can easily added back and renamed to
something like /usr/bin/go-xxhsum.

[1] https://github.com/Cyan4973/xxHash

BUG=b:187795288
TEST=Local builds; CQ.

Change-Id: I02a255582b4641ba9ab6dbcbdf26f60204c0b05b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3576436
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>