dev-lang/rust: upgrade to Rust 1.60.0
This upgrades Rust to 1.60.0 from 1.58.1.

Since instrument-coverage is now stable, this changes
cros-rust.eclass to pass -Cinstrument-coverage instead of
-Zinstrument-coverage. Except we actually have instrument
coverage commented out for Rust (b:230127632), so we keep
it that way.

To avoid b:229824406 (relocation refers to discarded section),
this change includes a patch which fixes the issue by causing
-Cpasses= to only affect pre-link stages, not link-time
optimization. The upstream pull request for this fix is
https://github.com/rust-lang/rust/pull/97969, and has
been merged into upstream rust.

Also added the patch

rust-1.60.0-Revert-DebugInfo-Re-enable-instruction-referencing-f.patch

to fix a segfault that appears when doing `sudo emerge-betty chunnel`
when Betty is built with a fuzzer profile.

Also install Rust sources, which will be needed by
rust-analyzer (see b:235120448), and disable a few
shellcheck warnings.

BUG=b:228448006, b:221426668, b:235120448
TEST=Use CQ to test the new Rust version

Change-Id: If1d613365baba35c081e0f928f7243e3e7525400
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3755890
Reviewed-by: Bob Haarman <inglorion@chromium.org>
Tested-by: Michael Benfield <mbenfield@google.com>