dev-go/gopsutil: fix installation path
CrOS is shipping and using gopsutil v3 but it does not
respect the Go modules path requirement for this specific
package. Gopsutil is the only known offender.

The requirement is for the major version (v3) to be added
as suffix to the path. The reason why this worked before
is because modules mode was disabled and GOPATH mode did
not enforce this requirement.

A separate CL is created for the gopsutil user tast-cmd
which should land at the same time as this to avoid
breakages.

BUG=b:187795288
TEST=Local builds; CQ.

Cq-Depend: chromium:3578762, chromium:3578808
Change-Id: I80691c7bd1edb170a340e2439a22cd1780408d47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3581731
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Auto-Submit: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Tested-by: Denis Nikitin <denik@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>