brya: Add osiris coreboot configuration
Create a new coreboot configuration for the osiris variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:229352299
TEST=FW_NAME=osiris emerge-brya coreboot chromeos-bootimage
Ensure that image-osiris.*.bin are created

Change-Id: I1a1a4341f4ede8e91b8969780a0dea20e05e26f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3588285
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: David Wu <david_wu@quanta.corp-partner.google.com>