tast-use-flags: Add `vkms` USE flag for tast.
This flag allows us to target tests that use vkms, the kernel's virtual
display interface. This is useful, for instance, for tests that exercise
DRM but don't necessarily want to provision a display in the lab.

Also had to add in some linter compliance bits. I am bad at shell, so if
you have a way to not disable the disable, I'm all ears. I tried read as
suggested by the linter and it did not go well.

BUG=b:229403705
TEST=Emerged the ebuild and it resulted in expected output

Change-Id: I58b5474ae1544ed4f455553eeb1ee02ac71f8487
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3591613
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Jim Shargo <jshargo@chromium.org>
Tested-by: Jim Shargo <jshargo@chromium.org>