llvm: asan to log to /tmp/asan
Currently the address sanitizer builds are failing,
potentially due to lack of permissions to write to the
original /var/log/asan/asan file. The image fails to
fully boot, and we have been unable to get an SSH connection
to view these logs.

This change has been shown in tryjobs to resolve the issue,
even if hacky. With this change applied, we should be able
to debug the builders further.

BUG=b:223869781
TEST=CQ
TEST=cros tryjob -g <this_cl> --latest-toolchain \
     amd64-generic-asan-tryjob

Change-Id: Id57d00e916d795b71d5f98aa60e12e6a749da2e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3590143
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Christopher Di Bella <cjdb@google.com>