bluez: Adding MaxController configuration
This change adds a MaxController=1 configuration to remove the ambiguity
situation that exists if the customer plugs in an external USB
controller.  It is still possible for the driver loading of both
adapters to race against each other, but is likely good enough for most
case since internal controllers are more likely to load first.

BUG=b:229747846
TEST=emerge-zork bluez and try to plug in an external USB controller and
     validate the MaxController limit worked as expected

Change-Id: If1ab847be3aedabcdeded48aad6b93e9b5857e5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3600636
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Tested-by: Alain Michaud <alainm@chromium.org>
Commit-Queue: Alain Michaud <alainm@chromium.org>