Revert "Reland "mesa-reven: build vaapi drivers for AMD and NVidia""
Some AMD hardware is showing discoloration when using VAAPI video
decoding. As a stopgap fix for release we are temporarily disabling
VAAPI on Nouveau and AMD until the root cause of the AMD rendering bug
can be fixed.

VAAPI was originally enabled on reven as a workaround for a camera crash
that seems to have been resolved.

This reverts commit 691e2e2b6bc2b0b584c90abaa2454b5c404e3344.

BUG=b:228101914
TEST=emerge-reven mesa-reven
TEST=equery-reven f mesa-reven | grep va/drivers doesn't show nouveau
and radeon drivers
TEST=Verify mesa-reven does not depend on libva with "equery-reven d
libva"

Change-Id: I839ed8535fb7ee0c85bd0c89a5b3b8cd19f09d83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3613761
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Michael Pobega <pobega@google.com>
Tested-by: Michael Pobega <pobega@google.com>