Move RDEPENDs out of post-cross
With the creation of dev-lang/rust-host, host ebuilds that depend on
Rust can be built before the cross compilers are installed which allows
them to be built or upgraded from source as opposed to being forced as
binary package only. This makes it possible to upgrade transitive
dependencies easier since the old version doesn't need to be kept around
for the binary packages not to break.

BUG=b:227239715
TEST=CQ passes

Change-Id: I2a4904e12a0977eb6339826528df322738e56dd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3624862
Commit-Queue: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>