cros-firmware: Refactor src_compile()
The current code of cros-firmware_src_compile() is written in a way that
pack_firmware.py calls are spread in 2 places. This makes the flow
control unnecessarily complicated. For example, if we want to do
something unconditionally, then we have no choice but to do it before
the first pack_firmware.py call.

Because the pack_firmware.py script will be called *only once* in any
case, refactor the code into a single if-else block. This improves
readability, and allows better flow control.

BUG=b:230679721
TEST=emerge-cherry chromeos-firmware-cherry
TEST=USE=tot_firmware emerge-cherry chromeos-firmware-cherry
TEST=USE=local_firmware emerge-cherry chromeos-firmware-cherry

Change-Id: Idcf9a2cf97c40aa50af9afd57793e9022e95873c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3625924
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Tested-by: Hung-Te Lin <hungte@chromium.org>