cros-firmware: Install signer_config.csv
Firmware provisioning in the lab sometimes wants to install a firmware
that is not pinned for chromeos-firmwareupdate yet. It does have access
to the firmware archive, but doesn't know which image from that archive
to use. The current workaround is to first try to look for
image-${MODEL}.bin, and if it's absent, image-${BOARD}.bin will be
tried. This approach is unstable and error prone, and sometimes leads to
nasty hacks such as CL:3555584.

To solve the problem, create a signer_config.csv and install it to
firmware/. Example of singer_config.csv:

 model_name,firmware_image,key_id,ec_image
 cherry,image-cherry.bin,DEFAULT,cherry/ec.bin
 dojo,image-dojo.bin,DEFAULT,dojo/ec.bin
 tomato,image-tomato.bin,CHERRYA,tomato/ec.bin

BUG=b:230679721
TEST=emerge-cherry chromeos-firmware-cherry
TEST=USE=local_firmware emerge-cherry chromeos-firmware-cherry

Cq-Depend: chromium:3616274
Change-Id: Idae10bfd46e5c965749f717331a3b433119fa31f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3625925
Reviewed-by: Bob Moragues <moragues@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>