lxd-4.0.9: use correct bin dir on non-amd64 arches.
On amd64, binaries end up in go/bin/...
On arm, binaries end up in go/bin/linux_arm/...

Pass the correct path to doexe, or otherwise we don't install
these binaries.

This caused a crash for the LXD4 experiment running LXD3 binaries
on ARM boards because the LXD4 ones were missing.

This CL is currently a no-op because we rolled back to lxd-4.0.7
to fix crbug.com/1321396

TEST=ebuild-tael lxd-4.0.9.ebuild install

Change-Id: I69855da44d641883fde1bed441774caf6367120c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3631688
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Emil Mikulic <easy@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Tested-by: Emil Mikulic <easy@google.com>
Reviewed-by: David Munro <davidmunro@google.com>