dev-embedded/hps-sdk: also build GCC 10.2 toolchain for RISC-V
The hps-firmware ebuild currently builds a private copy of
binutils+newlib+GCC for RISC-V, in order to compile tflite-micro used in
the FPGA ROM. But building all that and discarding it every time in
hps-firmware is slow and wasteful. Put it here in hps-sdk instead so
that it's only done once and then reused.

This uses the same sources as we currently have in hps-firmware, which
are from https://github.com/riscv-collab/riscv-gnu-toolchain, release
version 2021.04.23. Unfortunately we can't use the Makefile from
riscv-gnu-toolchain as is, because it assumes DESTDIR and --prefix are
the same, and that it can immediately install each package as it goes.
That assumption is true when installing a private copy into a temporary
directory but it doesn't fit the way Portage packages work, where they
are configured with a prefix and then installed under ${D}.

The problem is that building the toolchain is a 4 step process and each
step assumes the previous step has been installed into its prefix
already:

1. binutils
2. GCC "stage 1" (subset of the compiler without target libraries)
3. newlib
4. GCC "stage 2" (complete compiler with target libraries)

The Gentoo solution to this is crossdev, which builds and then installs
each of the 4 packages in sequence. Here we use a slightly more
convoluted approach to avoid splitting out 4 packages which need to be
managed by crossdev. We configure each package the same way that the
riscv-gnu-toolchain Makefile does, but with the prefix and sysroot
options tweaked. The first 3 steps are built and installed into
a temporary prefix under ${WORKDIR}, and we use that to compile the
final GCC with the real prefix. Then everything gets installed under
${D} as normal.

BUG=b:219591591
TEST=emerge hps-sdk &&
     cd ~/chromiumos/src/platform/hps-firmware2 &&
     PATH=/opt/hps-sdk/bin:$PATH scripts/build-fpga-rom-dev

Change-Id: I8c5485b3ffa4c6a789bfac461489130f93fb4403
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3640902
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>