vm_host_tools: Simplify vmlog_forwarder's stop condition
vmlog forwarder currently has a compound start and stop condition
for wilco devices. However, because the start condition is A or B and
the stop condition is A and B, there are several conditions (eg. only one
of A/B are started) where the service will never be stopped. Instead,
on wilco devices, use system-services as the stopping criteria.

BUG=b:202450969
TEST=CQ

Disallow-Recycled-Builds: test-failures
Change-Id: I7e5ce459443afb3fe02cc510071d5769970ff292
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3646609
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>