quipper: Add quipper patch for perf-5.15
This is a temporary solution to unblock perf-5.15 upgrade.
The patch is pulled from
https://github.com/google/perf_data_converter/commit/1019cb233dd08496574d36f1c86e4f2983253e0f.

The tests with perf.data.branch-4.14 are disabled since the testdata
binaries have to be updated to match with the new proto format but
binary patches are not supported by the patch manager.
This will be fixed with the next quipper upgrade.

BUG=b:223842077
TEST=unittests passed in github

Change-Id: Ic8799d9194ddf265bf6ee48971a0e25497c541d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3655928
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>
Tested-by: Denis Nikitin <denik@chromium.org>