brask: Add kuldax coreboot configuration
Create a new coreboot configuration for the kuldax variant
of the brask reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:233380254
TEST=FW_NAME=kuldax emerge-brask coreboot chromeos-bootimage
Ensure that image-kuldax.*.bin are created

Change-Id: I45686917a0b12831008782c6c9e117453e87bc10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3658912
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Queue: David Wu <david_wu@quanta.corp-partner.google.com>
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>