floss: Update minijail0 params for btmanagerd
btmanagerd now needs to access socket(PF_BLUETOOTH, ...) so we need to
modify the minijail0 params to provide this access.

We remove '-e' since this seems to prevent us from accessing the BT raw
socket and we add '-c 3400' to add the capabilities to access the raw
socket.

BUG=b:226644782
TEST=Build + deploy to device and verify btmanagerd works; ran
     test_that bluetooth_AdapterCLHealth.all_floss

Change-Id: I549d6865cc51471e6ffd6c3cf7b9919275626a3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3671486
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>