chromium-os-dev: Put hdctools behind a USE flag.
Adding hdctools was added back 10 years ago or so to make it convienent
to use any chrome device with a test image as your servo host.

This is rarely used and the build effort being expended to build it into
the test image for every device is not worth the benefit.

SIE's and the Power team have been consulted and agree to this change.

Any teams using this can just enable the USE and build a custom test
image.   Running hdctools on non validated chromeos platforms is not
supported by the HW Tools team.

BUG=b:234511803
TEST=build with/without use flag and check for servod in the /build

Change-Id: Ida7b7e0d54ec7ec85b69ce952930745fa0d21205
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3681803
Reviewed-by: Udaya Kiran Ammu <udaykiran@google.com>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Auto-Submit: Keith Haddow <haddowk@chromium.org>
Commit-Queue: Keith Haddow <haddowk@chromium.org>