chromeos-chrome: filter out sanitizer flags
Chrome has its own story for the use of asan/msan/ubsan that's
independent of ours. As we move toward LaCrOS, this will be increasingly
the case.

Because of this, filter out all sanitizer flags here. This makes the
application of checks in b/231357370 easier to reason about.

BUG=b:231357370
TEST=`emerge-atlas chromeos-chrome` with/without global sanitizer flags.

Change-Id: Ia84be0db46f17144d87423d9115af678016b55fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3692171
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>