dedede: Add shotzo coreboot configuration
Create a new coreboot configuration for the shotzo variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:235303242
TEST=FW_NAME=shotzo emerge-dedede coreboot chromeos-bootimage
Ensure that image-shotzo.*.bin are created

Change-Id: I76d995ffce7c1cf0b32bcd24ce04b9b2a105356a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3689144
Tested-by: Tony Huang <tony-huang@quanta.corp-partner.google.com>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Tony Huang <tony-huang@quanta.corp-partner.google.com>