sys-libs/compiler-rt: Pass -mfpu to compiler in addition to assembler
When linking EC, a GOT section was being generated. It was coming from
compiler-rt:

armv7m-cros-eabi-readelf --relocs \
  /usr/lib64/clang/14.0.0/lib/baremetal/libclang_rt.builtins-armv7m.a |\
  grep GOT

0000000c  00000760 R_ARM_GOT_PREL    00000000   __arm_fe_default_rmode

Looking at the compiler-rt code, __arm_fe_default_rmode only exists when
__ARM_FP is not defined. __ARM_FP should be defined by the compiler when
hardware floating point is enabled
https://developer.arm.com/documentation/dui0774/g/chr1383660321827.

Before this change, the flag was only passed to the assembler "-Wa".
Removing the "-Wa" so that the flag is passed to the compiler as well
fixes the issue.

BUG=b:234507656, b:234181908
TEST=sudo emerge cross-armv7m-cros-eabi/compiler-rt

Change-Id: Id5c2f98a53fdc4dc12e88cee6d11d9924c82fc50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3700723
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>
Tested-by: Tom Hughes <tomhughes@chromium.org>