platform.eclass: Add a script to update compdb to work outside chroot
Compilation database tells clangd how to compile the target C++ files.
Since the compilation database is generated inside chroot, we need to
make some modifications to it for it to be useful outside chroot.
This modification has been done in platform.eclass with sed and jq, but
it didn't have any tests.
In this patch we extract the logic to Python script and add unit tests
so that we can fix bugs without introducing regression. platform.eclass
is updated to just call the script. This should be safe because the
functionality is only used in local environment.

BUG=b:235426819
TEST=See Cq-Depend

Disallow-Recycled-Builds: novato-cq, brya-manatee-cq
Cq-Depend: chromium:3704981
Change-Id: I76f9b832e714031934e3a651038f3f8347c3e906
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3705104
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Keigo Oka <oka@chromium.org>
Tested-by: Keigo Oka <oka@chromium.org>