sys-devel/autofdo: backport patch to fix ambigous max arg failure
Starting with glibc >= 2.34, SIGSTKSZ is no longer a constant,
but is calculated resulting a value of type long. This causes
build failures because max expects args to be of same types.

Backport an upstream patch from third_party/abseil to fix it,
when upgrading this patch can be dropped.

BUG=b:235960683
TEST=Local build with glibc >= 2.34.

Change-Id: Ib0588e7e8ff65630f8f748ffd84a80d53355361e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3702289
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>