nissa: Add joxer coreboot configuration
Create a new coreboot configuration for the joxer variant
of the nissa reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

BUG=b:236086879
TEST=FW_NAME=joxer emerge-nissa coreboot chromeos-bootimage
Ensure that image-joxer.*.bin are created

Change-Id: I3baf2e042c59b5238f06dde0cd0985a6f0ed7e32
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3706946
Tested-by: Mark Hsieh <mark_hsieh@wistron.corp-partner.google.com>
Commit-Queue: Mark Hsieh <mark_hsieh@wistron.corp-partner.google.com>
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>