chromeos-base/crosvm: add rand-0.8 dependency
We already use a 0.8 version (via a rand = "*" in Cargo.toml), but we
didn't declare it in the ebuild. Add it to be sure it is available
before building crosvm.

For now, keep rand-0.6 as well; we can clean this up later once all
crosvm rand uses in Cargo.toml have been updated to 0.8.

BUG=b:236978141
TEST=emerge-hatch crosvm

Change-Id: Ic717f4412750484c2c4cb5b9fc7f239cfbee3f89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3722127
Auto-Submit: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>