sirenia: Allow `trichechus` on rootfs in ManaTEE
Currently `trichechus` is included on the root filesystem on builds
when `USE=manatee` is not present, but only available in the in
the initramfs on `USE=manatee` builds.

Having the option of including `trichechus` in the rootfs --- even on
ManaTEE boards --- is useful for development, such as allowing ManaTEE
developers and TEE app authors to run the stack in CrOS userspace.

Update the ebuild to always include `trichechus` when `USE=sirenia`
is set, independent of whether `USE=manatee` is enabled or not.

BUG=b:236892700
TEST=```
  sudo emerge sirenia
  sudo USE=sirenia emerge sirenia
  emerge-$BOARD-manatee sirenia && cros deploy $TARGET sirenia
  USE=sirenia emerge-$BOARD-manatee sirenia && cros deploy $TARGET sirenia
```

Change-Id: I6b29b696534eb78d5d90d2a364289e2cd7df3da7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3739323
Tested-by: David Greenaway <dgreenaway@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>
Auto-Submit: David Greenaway <dgreenaway@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>