cros-config-bsp.eclass now generates project-specific joined.jsonproto.
The ebuild process now generates joined.jsonproto based on the
checked-in imported.jsonproto (produced by the backfiller) and
config.jsonproto (produced by the ebuild process). The project-specific
joined.jsonproto files are installed in the /tmp dir where they will be
accessible to recipes via the Build API as artifacts.

BUG=b:195298103
TEST=performed a json comparison between the checked-in joined.jsonproto
and the produced version. Verified that the correct files were available
in the tmp dir of the sysroot for the board.

Change-Id: Icb6fab4a760471a7c2bc41ef19f2e5c146625f39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3599967
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Gilberto Contreras <gcontreras@google.com>
Tested-by: Gilberto Contreras <gcontreras@google.com>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>