chromeos-base/hps-sign-rom: split out from hps-firmware-tools
Introduce a dedicated ebuild just for the hps-sign-rom utility, which
signs HPS firmware images.

Also add it to the SDK target, so that we can grab the resulting binary
from the SDK prebuilts and use it on the signer.

This adds 822KiB to the size on disk of the SDK:

 * chromeos-base/hps-sign-rom-9999
         Total files : 3
         Total size  : 822.91 KiB

BUG=b:232884923
TEST=emerge-brya hps-firmware  # it also emerges hps-sign-rom host pkg

Change-Id: I1a7c1f78fc01af90ab9841c49c49aa9a1e51c577
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3750683
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>