sys-kernel/linux-headers: backport sys_resq & bump minijail
Newer glibc start making use of the restartable sequence
syscall to improve performance so we need to build the
CrOS userspace to allow it on kernels newer than 4.18.

Besides backporting the definitions to linux-headers, we
also bump minijail to cause a rebuild against the new
headers, to allow the glibc syscall existence probes
(check for -ENOSYS) even on kernels not supporting rseq.

BUG=b:235960683
TEST=Local builds against glibc 2.35

Disallow-Recycled-Builds: test-failures
Change-Id: I80706a5e99260261d45205a76609e8a4600a402b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3749266
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>