Enable display_LidCloseOpen chameleon test
The test display_LidCloseOpen has been in the repo, but was not
running. The most likely culprit for this is that the test has not
been added to the ebuild file. Fixing that now and hoping to see
test runs in the lab.

BUG=b:234759299
TEST=Presubmit checks

Change-Id: Id408a970e5e7113ec330422cd35d3af61d802c7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3761310
Tested-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: Vinayak Suley <vsuley@chromium.org>
Auto-Submit: Vinayak Suley <vsuley@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Tested-by: Vinayak Suley <vsuley@chromium.org>