Bluetooth: append BQR uuid to experimental per quality.conf file
In this patch, the BQR uuid is appended to the experimental command
option of bluetoothd only when the quality.conf file indicates that
BQR is enabled.

In the longer term, if we decide to enable BQR globally by default,
Just set

  experimental="${experimental},330859bc-7506-492d-9370-9a6f0614037f"

unconditionally in start_bluetoothd.sh.

BUG=b:242009925
TEST=Observe the log messages in /var/log/bluetooth.log.

Test 1: BQR is disabled initially.
[bluetooth.log] INFO bluetoothd: Set quality report adapter 0 enable 0

Test 2: BQR can be enabled by the dbus method as
$ dbus-send --system --print-reply --dest=org.bluez
  /org/chromium/Bluetooth org.chromium.Bluetooth.Debug.SetQuality byte:1
[bluetooth.log] INFO bluetoothd: Set quality report adapter 0 enable 1

Test 3: BQR will keep enabled when bluetoothd is restarted now.
$ restart bluetoothd
[bluetooth.log] INFO bluetoothd: Set quality report adapter 0 enable 1

Test 4: BQR can be disabled by the dbus method as
$ dbus-send --system --print-reply --dest=org.bluez
  /org/chromium/Bluetooth org.chromium.Bluetooth.Debug.SetQuality byte:0
[bluetooth.log] INFO bluetoothd: Set quality report adapter 0 enable 0

Test 5: BQR will keep disabled when bluetoothd is restarted now.
$ restart bluetoothd
[bluetooth.log] INFO bluetoothd: Set quality report adapter 0 enable 0

Test 6: If we append the BQR uuid to experimental unconditionally in
start_bluetoothd.sh, and do "restart bluetoothd", we will get
[bluetooth.log] INFO bluetoothd: Set quality report adapter 0 enable 1

Cq-Depend: chromium:3828287
Change-Id: I1a6cc5a980b0ee0af1d872b95976940d3ca3c7c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3827273
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>