kconfiglib: Move from portage-stable + patch
There is a customised copy of kconfiglib in platform/ec/util which is
the same as this package except it also includes one PR authored by sjg
that allows passing in extra directories to the Kconfig search path.

Add the patch so that the ec copy can be deleted.

BUG=b:238773780
TEST=sudo emerge dev-python/kconfiglib
TEST=cd platform/ec ; rm util/kconfiglib.py ; util/run_tests.sh

Change-Id: I26888c522549ee88e935a4b15b4caa27c24e46dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3763390
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>