coreboot/coreboot-9999: add USE flag to use private FSP headers
Right now there is no option to use FSP headers which is built with
FSP without changing .config file. This patch add USE flag which
gives a option to select FSP headers which is built with FSP while
building a coreboot.
This option is necessary for use with the builders by passing a flag instead of editing the config files

Change-Id: Ib6181ff7310a9509259361a5a872e91f49193f89
Signed-off-by: Ronak Kanabar <ronak.kanabar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3767413
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>