media-libs/clvk: add patch to keep testing against opencl1.2
Last uprev switched clvk default to opencl3.0.
While we are interested to activate it upstream to start looking at
how clvk is behaving against opencl3.0, we want to keep clvk in
chromeos in opencl1.2 in order to know which tests are still to be
fixed to get to opencl1.2.

Some tests are failing when targeting opencl3.0 while they are passing
opencl1.2.
We want those test to be considered are pass for now are they are no
more work needed for them to get to the opencl1.2 compliance.

BUG=b:239318871,b:238710823
TEST=emerge-volteer clvk
TEST=tast run ${dut} graphics.OpenclCts.${tests}

Change-Id: I847bcc6f90d7a6598464bc590e218ef1907937aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3770109
Auto-Submit: Romaric Jodin <rjodin@chromium.org>
Commit-Queue: Ilja Friedel <ihf@chromium.org>
Tested-by: Romaric Jodin <rjodin@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>