hplip: Fix fuzzer bug
The fuzzer data was getting read in as garbage data in the header,
so a very very large page was trying to get written, which caused
the fuzzer test to time out.

Added some checks to try and determine the printers max page size
and don't print pages larger than that size.

BUG=b:235635595
TEST=hpcups_fuzzer clusterfuzz-testcase-minimized-hpcups_fuzzer-4968817337171968

Change-Id: I55ee31c55518275ba38fa0c4728e0cf07bc5c9e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3780889
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Nathan Muggli <nmuggli@google.com>
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Commit-Queue: Nathan Muggli <nmuggli@google.com>