sci-electronics/yosys-f4pga-plugins: new package for HPS
This package provides a plugin, dsp-ff, which provides a useful
optimisation for HPS gateware synthesis. It can pack flip-flops into DSP
elements on Crosslink-NX.

We have to introduce a subslot in the yosys ebuild, so that
yosys-f4pga-plugins can depend on the exact subslot. Because the plugins
are linked against internal bits of yosys, they are expected to break
any time anything changes in yosys, so they always need to rebuilt.

This adds another 334KiB to the size on disk of the SDK:

 * sci-electronics/yosys-f4pga-plugins-1.0.0_p20220720
         Total files : 7
         Total size  : 334.04 KiB

BUG=b:239654663
TEST=sudo emerge yosys-f4pga-plugins

Change-Id: I69ddc32596d6b6b399cec958bd13e5aba4231f0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3779022
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>