Clean up remains of lvmd and lvmd-client.
This CL is a follow up to CL:3779024. That CL reverted CL:3736204 but
this wasn't enough to fix annealing. The reason seems to be that in
between CL:3736204 and CL:3779024, we got
1f105bdc5598090a52931d60d30e012c364e31ed [1] which upreved lvmd and
lvmd-client. Therefore, even after CL:3779024, those upreved files
remain and annealing fails trying to find the corresponding -9999
ebuilds.

This CL just deletes the lvmd-client and lvmd folders to effectively
finish the revert of CL:3736204.

[1] https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/1f105bdc5598090a52931d60d30e012c364e31ed

BUG=b:236007986
TEST=None

Change-Id: I16699bada25e3983481bbb268e577f142366e871
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3780887
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Andres Calderon Jaramillo <andrescj@google.com>
Commit-Queue: Andres Calderon Jaramillo <andrescj@google.com>