chromeos-base/hiberman: Start when trunksd does
The kernel-based hibernate encryption work at crrev.com/c/3603369 allows
the kernel to verify the authenticity and integrity of the hibernate
image on its own. To do this, it interacts directly with the TPM, doing
sealing and unsealing operations. In order to do this it needs a TPM
context handle, of which there are only 3. Normally trunksd manages
these handles, and indeed when we go to resume in an encrypted world we
find trunksd squatting on all the loaded context handles in the TPM.

The hibernate resume process only needs this handle for a moment at the
very beginning of resume to do the initial decryption of the hibernate
key. We want this process to start as early in boot as possible anyway,
to get through the long I/O ASAP. Make sure hiberman resume starts
before trunksd so that hiberman gets a chance to use this TPM handle
before trunksd consumes it.

BUG=b:237086416
TEST=Resume from hibernate with kernel-based encryption

Change-Id: I4295b4c6d22a58b55c976f0d6bda74fcb6c885d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3787560
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>