bpftool: add package
Add package for bpftool. This package differs from the upstream version
in that it depends on the github mirror rather than pulling down kernel
source and patching it.

Advantage of this is that the mirror has a cleaner versioning scheme and
does not have python dependencies for document generation.

This tool is used to generate eBPF skeleton helpers that are useful for
loading, attaching and communicating with eBPF applications.

BUG=b:239081075

TEST=emerged on amd64-generic, made as a dependency of a package that
uses ebpf and verifies that the toolchain work.

Change-Id: I1da6e1b0e52513260a4a99add823932be8e38592
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3824581
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jason Ling <jasonling@chromium.org>
Tested-by: Jason Ling <jasonling@chromium.org>