cros-kernel2: Drop CONFIG_DEBUG_INFO_DWARF4 for kgdb builds
CONFIG_DEBUG_INFO_DWARF4 is already the default for chromeos-4.4 to
chromeos-5.10, so there is no need to select it explicitly. For
chromeos-5.15 and later, the default is now DWARF5. USE=kgdb
overwrites it with CONFIG_DEBUG_INFO_DWARF4, which is undesirable.
Remove it.

BUG=b:240333352
TEST=Build with USE="kgdb vtconsole" ...

Change-Id: I2c1c9a6047dea99e76d07daf614600b4d36641ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3790462
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>