ureadahead: Try /sys/kernel/tracing first
In recent kernels, tracefs is a separate filesystem with a separate
mountpoint, to avoid relying on debugfs. Try that first, and fall back
to the debugfs path if we have trouble.

This will allow debugfs to be optional, especially in early boot.

Add an 'oom score' to silence linters while we're at it.

BUG=b:240487520
TEST=ureadahead --dump

Change-Id: I484f8a030a9ee36a66cb8160c8beb6c8cae72af0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3791879
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>