trunks: Wait for hiberman resume init
In the kernel-hardened hibernate design, the kernel verifies the
integrity of the hibernate image on its own by doing seal/unseal
operations directly with the TPM. For this, the kernel needs a TPM
handle. Normally trunksd manages the juggling act of loading and
unloading TPM handles, but it can't provide this service for the kernel
itself. So we need to make sure the hibernate mechanism is only
attempting to use TPM handles when trunksd is out of the way.

On suspend, this is fairly straightforward, as powerd has already
emitted its SuspendImminent signal, and upon receiving that trunksd
saves and flushes its handles out of the TPM, freeing them up for the
kernel to use.

But on resume, hiberman's access to load the kernel key happens early
during init, so that we can frontload as much of the expensive resume
work as is feasible. This races with trunksd starting and squatting on
all 3 of the TPM handles.

If USE=hibernate is set, delay starting trunksd until hiberman emits a
signal indicating it's done using the TPM. The start conditions for
hiberman are the same as for trunksd, and the signal can usually be
emitted right away, so there should be little to no impact on boot time.
Additionally, add an "or stopped hiberman" so that if hiberman crashes
before emitting the signal, trunksd will still start.

BUG=b:237086416
TEST=Resume from hibernate, stop seeing TPM errors sometimes

Cq-Depend: chromium:3803806
Change-Id: I7838470ae9462defe216f4763297b690784eb652
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3803825
Commit-Queue: Evan Green <evgreen@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>